The stream written by client is read by server socket.
