+++ src/main/java/ReadingSocket.java
@@ -1,8 +1,7 @@
-import java.io.BufferedReader;
 import java.net.Socket;
 
 public interface ReadingSocket {
 
     Socket acceptConnection();
-    BufferedReader getStreamFromClient(WritingSocket clientSocket);
+    String readStream(WritingSocket clientSocket);
 }
